runtime-sdk: Static call/query/message result handler dispatch
Refactors the way method dispatch is performed by removing the method
registries and replacing them with static dispatch in the form of a
MethodHandler trait which needs to be provided by modules.

Additionally the contexts are made more composable by making sure that
the we take Context as a generic type argument everywhere and improving
the way simulation mode is handled.

Each context is now also bound to a specific Runtime type to allow for
nested method dispatch.