ts-web: move tools to packages
It's nice to have dependencies like jest listed in the same
manifest where it's used.

I'm moving prettier too. This will make us overall less locked in
to having everything in this monorepo.

We now have some automation to bump our dependency versions, so
it's less of a burden to mention these in each package's
package.json.