runtime-sdk: Refactor state handling
The Context is now only used to serve read-only environment state
(e.g. information about the current block) and provide the Runtime type
for static dispatch. All mutable state (including storage) has been
moved into a new type, State, which supports nesting and transactions
(similar to the previous CurrentStore which only provided this for
storage).

This allows many things to be simplified and a better integration with
the EVM's substates (which now no longer require hacks to handle events
and messages).